services/workout-extract/requirements.txt
google-genai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
garminconnect>=0.2.38
garth>=0.4.0
python-dotenv>=1.0.0
